rename model module
