--- domain/src/main/java/com/onlydust/marketplace/indexer/domain/model/clean/CheckRun.java
@@ -1,4 +1,0 @@
-package com.onlydust.marketplace.indexer.domain.model.clean;
-
-public record CheckRun(Long id) {
-}
