--- domain/src/main/java/com/onlydust/marketplace/indexer/domain/model/clean/CodeReview.java
@@ -1,4 +1,0 @@
-package com.onlydust.marketplace.indexer.domain.model.clean;
-
-public record CodeReview(Long id, User author) {
-}
