--- domain/src/main/java/com/onlydust/marketplace/indexer/domain/model/clean/Commit.java
@@ -1,4 +1,0 @@
-package com.onlydust.marketplace.indexer.domain.model.clean;
-
-public record Commit(String sha, User author) {
-}
