--- domain/src/main/java/com/onlydust/marketplace/indexer/domain/model/clean/User.java
@@ -1,8 +1,0 @@
-package com.onlydust.marketplace.indexer.domain.model.clean;
-
-import com.onlydust.marketplace.indexer.domain.model.raw.RawSocialAccount;
-
-import java.util.List;
-
-public record User(Long id, String login, List<RawSocialAccount> socialAccounts) {
-}
