+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/clean/CheckRun.java
@@ -1,0 +1,4 @@
+package com.onlydust.marketplace.indexer.domain.models.clean;
+
+public record CheckRun(Long id) {
+}
