+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/clean/CodeReview.java
@@ -1,0 +1,4 @@
+package com.onlydust.marketplace.indexer.domain.models.clean;
+
+public record CodeReview(Long id, User author) {
+}
