+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/clean/Commit.java
@@ -1,0 +1,4 @@
+package com.onlydust.marketplace.indexer.domain.models.clean;
+
+public record Commit(String sha, User author) {
+}
