+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/clean/User.java
@@ -1,0 +1,8 @@
+package com.onlydust.marketplace.indexer.domain.models.clean;
+
+import com.onlydust.marketplace.indexer.domain.models.raw.RawSocialAccount;
+
+import java.util.List;
+
+public record User(Long id, String login, List<RawSocialAccount> socialAccounts) {
+}
