+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/raw/RawCommit.java
@@ -1,4 +1,4 @@
-package com.onlydust.marketplace.indexer.domain.model.raw;
+package com.onlydust.marketplace.indexer.domain.models.raw;
 
 
 import lombok.EqualsAndHashCode;
@@ -8,5 +8,8 @@ import lombok.Value;
 @Value
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor(force = true)
-public class RawIssue extends JsonDocument {
+public class RawCommit extends JsonDocument {
+    String sha;
+    RawUser author;
+    RawUser committer;
 }
