+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/raw/RawIssue.java
@@ -1,4 +1,4 @@
-package com.onlydust.marketplace.indexer.domain.model.raw;
+package com.onlydust.marketplace.indexer.domain.models.raw;
 
 
 import lombok.EqualsAndHashCode;
@@ -8,8 +8,5 @@ import lombok.Value;
 @Value
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor(force = true)
-public class RawRepo extends JsonDocument {
-    Long id;
-    String name;
-    RawUser owner;
+public class RawIssue extends JsonDocument {
 }
