+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/raw/RawRepo.java
@@ -1,4 +1,4 @@
-package com.onlydust.marketplace.indexer.domain.model.raw;
+package com.onlydust.marketplace.indexer.domain.models.raw;
 
 
 import lombok.EqualsAndHashCode;
@@ -8,7 +8,8 @@ import lombok.Value;
 @Value
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor(force = true)
-public class RawUser extends JsonDocument {
+public class RawRepo extends JsonDocument {
     Long id;
-    String login;
+    String name;
+    RawUser owner;
 }
