+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/services/IndexingService.java
@@ -1,7 +1,7 @@
 package com.onlydust.marketplace.indexer.domain.services;
 
 import com.onlydust.marketplace.indexer.domain.exception.NotFound;
-import com.onlydust.marketplace.indexer.domain.model.clean.*;
+import com.onlydust.marketplace.indexer.domain.models.clean.*;
 import com.onlydust.marketplace.indexer.domain.ports.out.RawStorageReader;
 import lombok.AllArgsConstructor;
